+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [3.0.7] - 2025-09-05
+
+### Added
+
+- Support for GET /users/{userId}/plans
+- Support for GET /2.0/users
+- Support for DELETE /users/{userId}/plans/{planId}
+
 ## [3.0.5] - 2025-04-07
 
 ### Changed